+++ .gitignore
@@ -1,2 +1,3 @@
 ./test.conf
 builds
+tmp
